Fix MoveToParent
In some circumstances, the presenter will present the sheet before the
function didMoveToParent hast been called, which is responsible to
determine the proxyPresenter. This will then cause the actual presenter
to be nil and the sheet will not be presented at all.

This fix ensures the proxyPresenter will be set in `willMoveToParent`
which will be called _before_ the presenter tries to get the actual
presenter and performs the presentation of the sheet.

Note: the code which is responsible to the determine the actual
presenter should be refactored in a later step in order to make it more
clean and tidy.